apps/api/requirements.txt
# Web framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pgvector==0.2.5

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Auth
pyjwt==2.8.0
cryptography==42.0.2
httpx==0.26.0

# OpenAI
openai==1.12.0

# Utilities
python-dotenv==1.0.1
structlog==24.1.0

# Error Tracking
sentry-sdk[fastapi]==1.40.0

# PDF Generation
fpdf2==2.7.8

# CSV Processing
pandas==2.2.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0

# Linting
ruff==0.2.1

